Move stuff to descriptor.rs
