+++ librazer/src/device.rs
@@ -1,51 +1,8 @@
-use crate::feature;
+use crate::descriptor::{Descriptor, SUPPORTED};
 use crate::packet::Packet;
 
 use anyhow::{anyhow, Context, Result};
 use std::{thread, time};
-
-// model_number_prefix shall conform to https://mysupport.razer.com/app/answers/detail/a_id/5481
-#[derive(Debug, Clone)]
-pub struct Descriptor {
-    pub model_number_prefix: &'static str,
-    pub name: &'static str,
-    pub pid: u16,
-    pub features: &'static [&'static str],
-}
-
-pub const SUPPORTED: &[Descriptor] = &[
-    Descriptor {
-        model_number_prefix: "RZ09-0483T",
-        name: "Razer Blade 16” (2023) Black",
-        pid: 0x029f,
-        features: &[
-            "battery-care",
-            "fan",
-            "kbd-backlight",
-            "lid-logo",
-            "lights-always-on",
-            "perf",
-        ],
-    },
-    Descriptor {
-        model_number_prefix: "RZ09-0482X",
-        name: "Razer Blade 14” (2023) Mercury",
-        pid: 0x029d,
-        features: &[
-            "battery-care",
-            "fan",
-            "kbd-backlight",
-            "lights-always-on",
-            "perf",
-        ],
-    },
-];
-
-const _VALIDATE_FEATURES: () = {
-    crate::const_for! { device in SUPPORTED => {
-        feature::validate_features(device.features);
-    }}
-};
 
 pub struct Device {
     device: hidapi::HidDevice,
